Enum cosmwasm_std::StdError

pub enum StdError {

    VerificationErr {
        source: VerificationError,
    },
    RecoverPubkeyErr {
        source: RecoverPubkeyError,
    },
    GenericErr {
        msg: String,
    },
    InvalidBase64 {
        msg: String,
    },
    InvalidDataSize {
        expected: u64,
        actual: u64,
    },
    InvalidHex {
        msg: String,
    },
    InvalidUtf8 {
        msg: String,
    },
    NotFound {
        kind: String,
    },
    ParseErr {
        target_type: String,
        msg: String,
    },
    SerializeErr {
        source_type: String,
        msg: String,
    },
    Overflow {
        source: OverflowError,
    },
    DivideByZero {
        source: DivideByZeroError,
    },
    ConversionOverflow {
        source: ConversionOverflowError,
    },

}

Structured error type for init, execute and query.

This can be serialized and passed over the Wasm/VM boundary, which allows us to use structured error types in e.g. integration tests. In that process backtraces are stripped off.

The prefix “Std” means “the standard error within the standard library”. This is not the only result/error type in cosmwasm-std.

When new cases are added, they should describe the problem rather than what was attempted (e.g. InvalidBase64 is preferred over Base64DecodingErr). In the long run this allows us to get rid of the duplication in “StdError::FooErr”.

Checklist for adding a new error:

Add enum case

Add creator function in std_error_helpers.rs

Variants

VerificationErr

Fields

source: VerificationError

RecoverPubkeyErr

Fields

source: RecoverPubkeyError

GenericErr

Fields

msg: String

Whenever there is no specific error type available

InvalidBase64

Fields

msg: String

InvalidDataSize

Fields

expected: u64

actual: u64

InvalidHex

Fields

msg: String

InvalidUtf8

Fields

msg: String

Whenever UTF-8 bytes cannot be decoded into a unicode string, e.g. in String::from_utf8 or str::from_utf8.

NotFound

Fields

kind: String

ParseErr

Fields

target_type: String

the target type that was attempted

msg: String

SerializeErr

Fields

source_type: String

the source type that was attempted

msg: String

Overflow

Fields

source: OverflowError

DivideByZero

Fields

source: DivideByZeroError

ConversionOverflow

Fields

source: ConversionOverflowError

Implementations

impl StdError

pub fn verification_err(source: VerificationError) -> Self

pub fn recover_pubkey_err(source: RecoverPubkeyError) -> Self

pub fn generic_err(msg: impl Into<String>) -> Self

pub fn invalid_base64(msg: impl ToString) -> Self

pub fn invalid_data_size(expected: usize, actual: usize) -> Self

pub fn invalid_hex(msg: impl ToString) -> Self

pub fn invalid_utf8(msg: impl ToString) -> Self

pub fn not_found(kind: impl Into<String>) -> Self

pub fn parse_err(target: impl Into<String>, msg: impl ToString) -> Self

pub fn serialize_err(source: impl Into<String>, msg: impl ToString) -> Self

pub fn overflow(source: OverflowError) -> Self

pub fn divide_by_zero(source: DivideByZeroError) -> Self