Struct cosmwasm_std::Decimal256

pub struct Decimal256(_);

A fixed-point decimal value with 18 fractional digits, i.e. Decimal256(1_000_000_000_000_000_000) == 1.0

The greatest possible value that can be represented is 115792089237316195423570985008687907853269984665640564039457.584007913129639935 (which is (2^256 - 1) / 10^18)

Implementations

impl Decimal256

pub const DECIMAL_PLACES: u32 = 18u32

The number of decimal places. Since decimal types are fixed-point rather than floating-point, this is a constant.

pub const MAX: Self = _

The largest value that can be represented by this decimal type.

pub const MIN: Self = _

The smallest value that can be represented by this decimal type.

pub const fn new(value: Uint256) -> Self

Creates a Decimal256 from Uint256 This is equivalent to Decimal256::from_atomics(value, 18) but usable in a const context.

pub const fn raw(value: u128) -> Self

Creates a Decimal256 from u128 This is equivalent to Decimal256::from_atomics(value, 18) but usable in a const context.

pub const fn one() -> Self

Create a 1.0 Decimal256

pub const fn zero() -> Self

Create a 0.0 Decimal256

pub fn percent(x: u64) -> Self

Convert x% into Decimal256

pub fn permille(x: u64) -> Self

Convert permille (x/1000) into Decimal256

pub fn from_atomics(
atomics: impl Into<Uint256>,
decimal_places: u32
) -> Result<Self, Decimal256RangeExceeded>

Creates a decimal from a number of atomic units and the number of decimal places. The inputs will be converted internally to form a decimal with 18 decimal places. So the input 123 and 2 will create the decimal 1.23.

Using 18 decimal places is slightly more efficient than other values as no internal conversion is necessary.

Examples

let a = Decimal256::from_atomics(1234u64, 3).unwrap();
assert_eq!(a.to_string(), "1.234");

let a = Decimal256::from_atomics(1234u128, 0).unwrap();
assert_eq!(a.to_string(), "1234");

let a = Decimal256::from_atomics(1u64, 18).unwrap();
assert_eq!(a.to_string(), "0.000000000000000001");

let a = Decimal256::from_atomics(Uint256::MAX, 18).unwrap();
assert_eq!(a, Decimal256::MAX);

pub fn from_ratio(
numerator: impl Into<Uint256>,
denominator: impl Into<Uint256>
) -> Self

Returns the ratio (numerator / denominator) as a Decimal256

pub fn checked_from_ratio(
numerator: impl Into<Uint256>,
denominator: impl Into<Uint256>
) -> Result<Self, CheckedFromRatioError>

Returns the ratio (numerator / denominator) as a Decimal256

pub const fn is_zero(&self) -> bool

pub const fn atomics(&self) -> Uint256

A decimal is an integer of atomic units plus a number that specifies the position of the decimal dot. So any decimal can be expressed as two numbers.

Examples

// Value with whole and fractional part
let a = Decimal256::from_str("1.234").unwrap();
assert_eq!(a.decimal_places(), 18);
assert_eq!(a.atomics(), Uint256::from(1234000000000000000u128));

// Smallest possible value
let b = Decimal256::from_str("0.000000000000000001").unwrap();
assert_eq!(b.decimal_places(), 18);
assert_eq!(b.atomics(), Uint256::from(1u128));

pub const fn decimal_places(&self) -> u32

The number of decimal places. This is a constant value for now but this could potentially change as the type evolves.

See also Decimal256::atomics().

pub fn floor(&self) -> Self

Rounds value down after decimal places.

pub fn ceil(&self) -> Self

Rounds value up after decimal places. Panics on overflow.

pub fn checked_ceil(&self) -> Result<Self, RoundUpOverflowError>

Rounds value up after decimal places. Returns OverflowError on overflow.

pub fn checked_add(self, other: Self) -> Result<Self, OverflowError>

pub fn checked_sub(self, other: Self) -> Result<Self, OverflowError>

pub fn checked_mul(self, other: Self) -> Result<Self, OverflowError>

Multiplies one Decimal256 by another, returning an OverflowError if an overflow occurred.

pub fn pow(self, exp: u32) -> Self

Raises a value to the power of exp, panics if an overflow occurred.

pub fn checked_pow(self, exp: u32) -> Result<Self, OverflowError>

Raises a value to the power of exp, returning an OverflowError if an overflow occurred.

pub fn checked_div(self, other: Self) -> Result<Self, CheckedFromRatioError>

pub fn checked_rem(self, other: Self) -> Result<Self, DivideByZeroError>

pub fn sqrt(&self) -> Self

Returns the approximate square root as a Decimal256.

This should not overflow or panic.

pub fn abs_diff(self, other: Self) -> Self

pub fn saturating_add(self, other: Self) -> Self

pub fn saturating_sub(self, other: Self) -> Self

pub fn saturating_mul(self, other: Self) -> Self

pub fn saturating_pow(self, exp: u32) -> Self