Module cosmwasm_std::testing

Structs

BankQuerier

MockApi

MockQuerier

MockQuerier holds an immutable table of bank balances and configurable handlers for Wasm queries and custom queries.

StakingQuerier

Constants

MOCK_CONTRACT_ADDR

Functions

digit_sum

mock_dependencies

Creates all external requirements that can be injected for unit tests.

mock_dependencies_with_balance

Creates all external requirements that can be injected for unit tests.

mock_dependencies_with_balances

Initializes the querier along with the mock_dependencies. Sets all balances provided (you must explicitly set contract balance if desired).

mock_env

Returns a default enviroment with height, time, chain_id, and contract address You can submit as is to most contracts, or modify height/time if you want to test for expiration.

mock_ibc_channel

Creates an IbcChannel for testing. You set a few key parameters for handshaking, If you want to set more, use this as a default and mutate other fields

mock_ibc_channel_close_confirm

Creates a IbcChannelCloseMsg::CloseConfirm for testing ibc_channel_close.

mock_ibc_channel_close_init

Creates a IbcChannelCloseMsg::CloseInit for testing ibc_channel_close.

mock_ibc_channel_connect_ack

Creates a IbcChannelConnectMsg::ConnectAck for testing ibc_channel_connect.

mock_ibc_channel_connect_confirm

Creates a IbcChannelConnectMsg::ConnectConfirm for testing ibc_channel_connect.

mock_ibc_channel_open_init

Creates a IbcChannelOpenMsg::OpenInit for testing ibc_channel_open.

mock_ibc_channel_open_try

Creates a IbcChannelOpenMsg::OpenTry for testing ibc_channel_open.

mock_ibc_packet_ack

Creates a IbcPacketAckMsg for testing ibc_packet_ack. You set a few key parameters that are often parsed. If you want to set more, use this as a default and mutate other fields. The difference from mock_ibc_packet_recv is if my_channel_id is src or dest.

mock_ibc_packet_recv

Creates a IbcPacketReceiveMsg for testing ibc_packet_receive. You set a few key parameters that are often parsed. If you want to set more, use this as a default and mutate other fields

mock_ibc_packet_timeout

Creates a IbcPacketTimeoutMsg for testing ibc_packet_timeout. You set a few key parameters that are often parsed. If you want to set more, use this as a default and mutate other fields. The difference from mock_ibc_packet_recv is if my_channel_id is src or dest./

mock_info

Just set sender and funds for the message. This is intended for use in test code only.

mock_wasmd_attr

Only for test code. This bypasses assertions in new, allowing us to create _* Attributes to simulate responses from the blockchain

riffle_shuffle

Performs a perfect shuffle (in shuffle)

Type Definitions

MockQuerierCustomHandlerResult

The same type as cosmwasm-std’s QuerierResult, but easier to reuse in cosmwasm-vm. It might diverge from QuerierResult at some point.

MockStorage